Add infrastructure for writing of the conversions
In order to enforce abstraction (and separates the list dealing from the
rest) and still respects the Norm, I had to do some ugly things with
f_lstmaparray. The entry of that function shall always be controlled,
since the function pointer controls the end of the loop, and shall know
when to stop, in order not to try to use a NULL pointer.