Add test developpement support in Makefile
Requieres one test file by source file if test is enabled (=a test src
directory is defined).
Compile and link the test (with the same flags as those of the target of
that dir) then run it. If it succeeds, update the corresponding watch
file with a touch command. That way, test are remade only if their
prerequisite have changed.