Fix fifo take elem when last elem is taken.
In the event where the last element of the list was taken, the fifo
structure referenced an invalid pointer in its member end_lst.
In order to avoid that bug, I have to obtain the last link of the list
in begin_lst member.
However, in order to not go over all the list every time this function
is invoked, I check if the returned content is the same that the one
which was contained in the end_lst link before the operation, and only
update the end_lst ref if it is.
That check might misleading (differents links, including the last, might
eventually be referencing the same content) but in most cases, it will
mitigate the cost of that fix.